one more cast error on non-Mac platforms

+++ display_list/display_list_vertices.cc
@@ -117,7 +117,7 @@ DlVertices::DlVertices(DlVertexMode mode,
       offset += bytes;
       return ret;
     } else {
-      return 0ul;
+      return size_t(0);
     }
   };
